WebDemo/requirements.txt
# Requirements needed when generating releases. See BUILD.rst for details.
invoke >= 0.20
rellu >= 0.6
docutils >= 0.14
robotframework >= 3.1.1
robotframework-seleniumlibrary >= 3.3.1
webdriver-manager
